Moved `load_configuration` file location
The `load_configuration` function was previously included as a
static method within `Survey`; however, this isn't necessary since
`load_configuration` never uses `self` as an argument. Consequently,
it has been moved to `EchoPro.utils.data_file_validation`.
