+++ EchoPro/utils/data_file_validation.py
@@ -1,5 +1,69 @@
 from pathlib import Path
 from openpyxl import load_workbook
+import yaml
+
+def load_configuration( init_config_path: Path , 
+                        survey_year_config_path: Path ):
+    """
+    Loads the biological, NASC, and stratification
+    data using parameters obtained from the configuration
+    files.
+
+    Parameters
+    ----------
+    init_config_path : pathlib.Path
+        A string specifying the path to the initialization YAML file
+    survey_year_config_path : pathlib.Path
+        A string specifying the path to the survey year YAML file
+
+    Notes
+    -----
+    This function parses the configuration files and incorporates them into
+    the Survey class object. This initializes the `config` attribute that 
+    becomes available for future reference and functions.
+    """
+    ### Validate configuration files
+    # Retreive the module directory to begin mapping the configuration file location
+    #current_directory = os.path.dirname(os.path.abspath(__file__))
+
+    # Build the full configuration file paths and verify they exist
+    config_files = [init_config_path, survey_year_config_path]
+    config_existence = [init_config_path.exists(), survey_year_config_path.exists()] 
+
+    # Error evaluation and print message (if applicable)
+    if not all(config_existence):
+        missing_config = [ files for files, exists in zip( config_files, config_existence ) if not exists ]
+        raise FileNotFoundError(f"The following configuration files do not exist: {missing_config}")
+
+    ### Read configuration files
+    # If configuration file existence is confirmed, proceed to reading in the actual files
+    ## !!! TODO: Incorporate a configuration file validator that enforces required variables and formatting
+    init_config_params = yaml.safe_load(init_config_path.read_text())
+    survey_year_config_params = yaml.safe_load(survey_year_config_path.read_text())        
+
+    # Validate that initialization and survey year configuration parameters do not intersect
+    config_intersect = set(init_config_params.keys()).intersection(set(survey_year_config_params.keys()))
+    
+    # Error evaluation, if applicable
+    if config_intersect:
+        raise RuntimeError(
+            f"The initialization and survey year configuration files comprise the following intersecting variables: {config_intersect}"
+        )
+
+    ### Format dictionary that will parameterize the `config` class attribute
+    # Join the initialization and survey year parameters into a single dictionary
+    config_to_add = { **init_config_params , **survey_year_config_params }
+    
+    # Amend length/age distribution locations within the configuration attribute
+    config_to_add[ 'biometrics' ] = { 
+        'bio_hake_len_bin': init_config_params[ 'bio_hake_len_bin' ] ,
+        'bio_hake_age_bin': init_config_params[ 'bio_hake_age_bin' ]
+    }
+    
+    del config_to_add['bio_hake_len_bin'] , config_to_add['bio_hake_age_bin']
+    
+    # Pass 'full_params' to the class instance
+    return config_to_add
 
 def validate_data_columns( file_name: Path ,
                            sheet_name: str ,
